+++ src/pages/planos/index.tsx
@@ -1,0 +1,44 @@
+import type { PageProps } from 'gatsby'
+import { graphql } from 'gatsby'
+import type { PlanosQuery } from '@generated/graphql'
+import { mark } from 'src/sdk/tests/mark'
+import SimpleText from 'src/components/sections/SimpleText/SimpleText'
+import Section from 'src/components/sections/Section'
+
+export type Props = PageProps<PlanosQuery>
+function Page(props: Props) {
+  const {
+    data: { allContentfulPlanosTextoSimples },
+  } = props
+
+  return (
+    <Section>
+      <SimpleText
+        textReceived={allContentfulPlanosTextoSimples}
+        className="text-banner-bottom"
+      />
+    </Section>
+  )
+}
+
+export const querySSG = graphql`
+  query Planos {
+    site {
+      siteMetadata {
+        title
+        description
+        titleTemplate
+      }
+    }
+    allContentfulPlanosTextoSimples {
+      nodes {
+        text {
+          text
+        }
+      }
+    }
+  }
+`
+
+Page.displayName = 'Page'
+export default mark(Page)